requirements.txt
asgiref==3.4.1
Django==3.2.5
greenlet==1.1.0
mysqlclient==2.0.3
numpy==1.21.1
pandas==1.3.1
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.22
sqlparse==0.4.1
